first version of new file source
