+++ quixstreams/sources/community/file/__init__.py
@@ -1,4 +1,6 @@
 # ruff: noqa: F403
-from .file import *
+from .azure import *
+from .compressions import *
 from .formats import *
-from .origins import *
+from .local import *
+from .s3 import *
