--- quixstreams/sources/community/file/origins/__init__.py
@@ -1,4 +1,0 @@
-# ruff: noqa: F403
-from .azure import *
-from .local import *
-from .s3 import *
